+++ app/src/main/java/org/adaway/helper/PreferenceHelper.java
@@ -96,6 +96,17 @@ public class PreferenceHelper {
         );
     }
 
+    public static boolean getUpdateCheckAppStartup(Context context) {
+        SharedPreferences prefs = context.getSharedPreferences(
+                Constants.PREFS_NAME,
+                Context.MODE_PRIVATE
+        );
+        return prefs.getBoolean(
+                context.getString(R.string.pref_update_check_app_startup_key),
+                context.getResources().getBoolean(R.bool.pref_update_check_app_startup_def)
+        );
+    }
+
     public static boolean getUpdateCheckAppDaily(Context context) {
         SharedPreferences prefs = context.getSharedPreferences(
                 Constants.PREFS_NAME,