+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 - Add new unique source parser with parallel processing
 - Add an option to disable app update check at startup
+- Fix crash with source file when SAF permission is removed
 
-Special thanks to andy356 and fusionneur for their bug report.
+Special thanks to andy356, fusionneur and sr1canskhsia for their bug report.
 
 ## [5.2.1] - 2021-01-06